requirements.txt
flask==1.0.2
kafka-python==1.4.5
elasticsearch==6.3.1
tweepy==3.7.0
injector==0.15.0
flask-restplus==0.12.1
flask-injector==0.11.0
pyspark==2.4.0
bitarray==0.8.3
mmh3==2.5.1
findspark==1.3.0
nltk==3.4.0